services/detect-text/pyproject.toml
[project]
name = "aegis-detect-text"
version = "0.1.0"
description = "Text drift + fairness detection worker for Aegis (MMD on embeddings)"
requires-python = ">=3.13"
dependencies = [
  "aegis-shared",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.28.0",
  "pydantic>=2.10.0",
  "numpy>=2.1.0",
]

[project.optional-dependencies]
# Only install when actually running detect-text in production / Colab.
# Tests use random embeddings (no model loading), so the heavy NLP stack
# is opt-in.
nlp = [
  "torch>=2.5.0",
  "sentence-transformers>=3.0.0",
]

[tool.uv.sources]
aegis-shared = { workspace = true }

[dependency-groups]
dev = [
  "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aegis_detect_text"]
